refactor: move env-loading helpers to internal/cli (#529 slice 2)
Second slice of #529: move the env-file loading + flag-to-env binding from
cmd/bintrail/envload.go into internal/cli so a second binary (bintrail-pg, #527)
can reuse them, and so the soon-to-move read commands can call them from cli.

- internal/cli/env.go: loadEnvFile/parseAndSetEnv (unexported) + exported
  BindCommandEnv and the EnvBindings table (+ EnvBinding type) + its test.
- cmd/bintrail/envload.go shrinks to a one-line bindCommandEnv wrapper over
  cli.BindCommandEnv, so all ~23 command init()s keep calling bindCommandEnv(cmd)
  unchanged (zero churn, no behavior change).
- config_test.go cross-check repointed to cli.EnvBindings; config.go comment
  updated.

Pure no-op refactor. Foundation for the status pilot (next slice). Part of #529.
